requirements.txt
jinja2==3.1.3
python-dotenv==1.0.0
openai==1.12.0
pyyaml==6.0.1
gitpython==3.1.42
